# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deepcuts API",
    "description": "An API to access music artist metadata and recommendations.",
    "version": "v0.0.0"
  },
  "paths": {
    "/albums/{artist_name}": {
      "get": {
        "tags": [
          "deepcuts"
        ],
        "summary": "Get Albums By Artist",
        "description": "Get albums by artist name.",
        "operationId": "deepcuts-get_albums_by_artist",
        "parameters": [
          {
            "name": "artist_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artist Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetArtistAlbumsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recommend": {
      "post": {
        "tags": [
          "deepcuts"
        ],
        "summary": "Recommend Albums From Albums",
        "description": "Recommend albums based on shared artists.",
        "operationId": "deepcuts-recommend_albums_from_albums",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_deepcuts-recommend_albums_from_albums"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRecommendationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Album": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "discogs_release_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Release Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "credit_artist_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Credit Artist Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "Album",
        "description": "Schema for an album."
      },
      "Body_deepcuts-recommend_albums_from_albums": {
        "properties": {
          "album_titles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Titles"
          },
          "artist_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Names"
          },
          "genres": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Genre"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genres"
          }
        },
        "type": "object",
        "required": [
          "album_titles",
          "artist_names"
        ],
        "title": "Body_deepcuts-recommend_albums_from_albums"
      },
      "Genre": {
        "type": "string",
        "enum": [
          "acoustic",
          "afrobeat",
          "alt-rock",
          "alternative",
          "ambient",
          "anime",
          "black-metal",
          "bluegrass",
          "blues",
          "bossanova",
          "brazil",
          "breakbeat",
          "british",
          "cantopop",
          "chicago-house",
          "children",
          "chill",
          "classical",
          "club",
          "comedy",
          "country",
          "dance",
          "dancehall",
          "death-metal",
          "deep-house",
          "detroit-techno",
          "disco",
          "disney",
          "drum-and-bass",
          "dub",
          "dubstep",
          "edm",
          "electro",
          "electronic",
          "emo",
          "folk",
          "forro",
          "french",
          "funk",
          "garage",
          "german",
          "gospel",
          "goth",
          "grindcore",
          "groove",
          "grunge",
          "guitar",
          "happy",
          "hard-rock",
          "hardcore",
          "hardstyle",
          "heavy-metal",
          "hip-hop",
          "holidays",
          "honky-tonk",
          "House",
          "idm",
          "indian",
          "indie",
          "indie-pop",
          "industrial",
          "iranian",
          "j-dance",
          "j-idol",
          "j-pop",
          "j-rock",
          "jazz",
          "k-pop",
          "kids",
          "latin",
          "latino",
          "malay",
          "mandopop",
          "metal",
          "metal-misc",
          "metalcore",
          "minimal-techno",
          "movies",
          "mpb",
          "new-age",
          "new-release",
          "opera",
          "pagode",
          "party",
          "philippines-opm",
          "piano",
          "pop",
          "pop-film",
          "post-dubstep",
          "power-pop",
          "progressive-house",
          "psych-rock",
          "punk",
          "punk-rock",
          "r-n-b",
          "rainy-day",
          "reggae",
          "reggaeton",
          "road-trip",
          "rock",
          "rock-n-roll",
          "rockabilly",
          "romance",
          "sad",
          "salsa",
          "samba",
          "sertanejo",
          "show-tunes",
          "singer-songwriter",
          "ska",
          "sleep",
          "songwriter",
          "SOul",
          "soundtracks",
          "spanish",
          "study",
          "summer",
          "swedish",
          "synth-pop",
          "tango",
          "techno",
          "trance",
          "trip-hop",
          "turkish",
          "work-out",
          "world-music"
        ],
        "title": "Genre",
        "description": "Available genres."
      },
      "GetArtistAlbumsResponse": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/Album"
            },
            "type": "array",
            "title": "Albums"
          }
        },
        "type": "object",
        "required": [
          "albums"
        ],
        "title": "GetArtistAlbumsResponse",
        "description": "Schema for an album_response."
      },
      "GetRecommendationResponse": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/Album"
            },
            "type": "array",
            "title": "Albums"
          }
        },
        "type": "object",
        "required": [
          "albums"
        ],
        "title": "GetRecommendationResponse",
        "description": "Schema for a recommendation response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}